Success

+++ GA_clustering_optimization/src/ga_kmeans/Point.java
@@ -1,26 +1,26 @@
 package ga_kmeans;
 
 public class Point {
-	private int x,y;
+	private double x,y;
 	
-	public Point(int x, int y){
+	public Point(double x, double y){
 		this.x = x;
 		this.y = y;
 	}
 	
-	public int getX(){
+	public double getX(){
 		return x;
 	}
 	
-	public void setX(int x){
+	public void setX(double x){
 		this.x = x;
 	}
 	
-	public int getY(){
+	public double getY(){
 		return y;
 	}
 	
-	public void setY(int y){
+	public void setY(double y){
 		this.y = y;
 	}
